api/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "telegram-inline-keyboard-builder"
version = "1.0.0"
description = "Library-agnostic inline keyboard builder for Telegram bots (Aiogram, Pyrogram, python-telegram-bot)."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "neoncraftx", email = "neoncraftx@gmail.com" }
]
requires-python = ">=3.8"

keywords = [
  "telegram",
  "telegram-bot",
  "inline-keyboard",
  "aiogram",
  "pyrogram",
  "python-telegram-bot",
  "builder",
  "adapter"
]

classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
  "Topic :: Communications :: Chat"
]


[project.urls]
Homepage = "https://github.com/neoncraftx/telegram-inline-keyboard-builder"
Repository = "https://github.com/neoncraftx/telegram-inline-keyboard-builder"
Issues = "https://github.com/neoncraftx/telegram-inline-keyboard-builder/issues"


# Optional dependencies for adapter
[project.optional-dependencies]
aiogram = ["aiogram>=3.0"]
pyrogram = ["pyrogram>=2.0"]
ptb = ["python-telegram-bot>=20.0"]

# Install all adapters at once
all = [
  "aiogram>=3.0",
  "pyrogram>=2.0",
  "python-telegram-bot>=20.0"
]


[tool.setuptools]
packages = [
  "telegram_inline_keyboard_builder",
  "telegram_inline_keyboard_builder.core",
  "telegram_inline_keyboard_builder.adapters",
  "telegram_inline_keyboard_builder.builders"
]